requirements.txt
echo "flask
web3
python-dotenv
requests" > requirements.txt
